Remove fields ``date_created`` and ``date_modified``
These fields are redundant for logical deletion.
Model itself named ``LogicalDeleteModel`` and seeing
this fields inside it are surprising.

Also, there are a few apps that provide similarly
functionality:

* https://github.com/django-extensions/django-extensions
  It has ``django_extensions.db.models.TimeStampedModel``
* https://github.com/carljm/django-model-utils
  It has ``model_utils.models.TimeStampedModel``

So if somebody started to use one of this app or some another,
and then wants to start using ``pinax-models`` with logical
delete feature, he will get confusing by duplicating fields